Fixed issue of custom requests not receiving workspace data
It's an issue of pass by value vs. pass by reference. When the workspace is passed to the request, I think it's done in the early stages, when the workspace is empty. So an empty workspace is copied in, and any values aren't updated correctly. This fixes that (I've tested it myself in VS Code, with some adjustments).

To get this to work with the VSCode extension though, you'll need to go to editor.ts and replace all of the document.uri.toString() calls with document.uri.toString().replace("%3A", ":"). For some reason, the document uri presents the colon as %3A (which differs from how every other document uses colons in the Language Server)